@@ -1,8 +1,8 @@
 import { combineEpics } from 'redux-observable';
 import { referenceServiceEpic } from './reference/referenceOperations';
-import { blotterServiceEpic } from './blotter/blotterOperations';
+import { blotterServiceEpic } from '../ui/blotter/blotterOperations';
 import { pricingServiceEpic } from './pricing/pricingOperations'
-import { analyticsServiceEpic } from './analytics/analyticsOperations';
+import { analyticsServiceEpic } from '../ui/analytics/analyticsOperations';
 import { compositeStatusServiceEpic } from './compositeStatusService/compositeStatusServiceOperations';
 
 export default combineEpics(
