@@ -1,10 +1,11 @@
 import React from 'react';
 import { SmartComponent } from 'esp-js-react';
-import {Modal, Chrome} from '../../common/components';
-import {FooterView} from '../../footer/views';
-import {WorkspaceRegionView} from '../../regions/views/workspace';
-import {SingleItemRegionView} from '../../regions/views/singleItem';
-import {SidebarRegionView} from '../../regions/views/sidebar';
+import { Chrome, Modal } from '../../common/components';
+import FooterContainer from '../../footer/FooterContainer';
+import SidebarRegionContainer from '../../sidebar/SidebarRegionContainer';
+import { WorkspaceRegionView } from '../../regions/views/workspace';
+import { BlotterContainer } from '../../blotter';
+import { RegionWrapper } from '../../../redux/regions'
 import './shell.scss';
 import '../../common/styles/_base.scss';
 import '../../common/styles/_fonts.scss';
@@ -27,7 +28,7 @@ export default class ShellView extends React.Component {
       <SmartComponent modelId={wellKnownModelIds.chromeModelId} view={Chrome}>
         <div>
           <div className='shell__splash'>
-            <span className='shell__splash-message'>{model.appVersion}<br />Loading...</span>
+            <span className='shell__splash-message'>{model.appVersion}<br/>Loading...</span>
           </div>
           <div className='shell__container'>
             <Modal shouldShow={model.sessionExpired} title='Session expired'>
@@ -40,13 +41,20 @@ export default class ShellView extends React.Component {
               </div>
             </Modal>
             <div className='shell_workspace_blotter'>
-              <SmartComponent className='shell__workspace' modelId={wellKnownModelIds.workspaceRegionModelId} view={WorkspaceRegionView} />
-              <SmartComponent className='shell__blotter' modelId={wellKnownModelIds.blotterRegionModelId} view={SingleItemRegionView} />
+              <SmartComponent className='shell__workspace' modelId={wellKnownModelIds.workspaceRegionModelId}
+                              view={WorkspaceRegionView}/>
+              <RegionWrapper region="blotter">
+                <div className='shell__blotter'>
+                  <BlotterContainer/>
+                </div>
+              </RegionWrapper>
             </div>
-            <SmartComponent className='shell__sidebar' modelId={wellKnownModelIds.sidebarRegionModelId} view={SidebarRegionView} />
+            <RegionWrapper region="analytics">
+              <SidebarRegionContainer/>
+            </RegionWrapper>
           </div>
           <div className='shell__footer'>
-            <SmartComponent modelId={wellKnownModelIds.footerModelId} view={FooterView} />
+            <FooterContainer/>
           </div>
         </div>
       </SmartComponent>
